Add AppointmentTime as parent class for AppointmentStartTime and AppointmentEndTime
